Update target screen when screens change
We want to target a specific screen or the top left screen. Our target
screen may not always be available, and the top left screen may change.
Update the target screen whenever the screen layout changes.

Just hiding and showing the window does not appear to work, so
just reconstruct the widget.